@@ -1,28 +1,31 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net8.0</TargetFramework>
+		<RootNamespace>PPTail</RootNamespace>
+		<LangVersion>latest</LangVersion>
+		<Nullable>enable</Nullable>
+	</PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.FxCopAnalyzers" Version="3.0.0">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.1.1" />
-  </ItemGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.CodeAnalysis.NetAnalyzers" Version="8.0.0">
+			<PrivateAssets>all</PrivateAssets>
+			<IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+		</PackageReference>
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="3.1.32" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\PPTail.Common\PPTail.Common.csproj" />
-    <ProjectReference Include="..\PPTail.Data.Ef\PPTail.Data.Ef.csproj" />
-    <ProjectReference Include="..\PPTail.Data.FileSystem\PPTail.Data.FileSystem.csproj" />
-    <ProjectReference Include="..\PPTail.Data.Forestry\PPTail.Data.Forestry.csproj" />
-    <ProjectReference Include="..\PPTail.Data.MediaBlog\PPTail.Data.MediaBlog.csproj" />
-    <ProjectReference Include="..\PPTail.Data.NativeJson\PPTail.Data.NativeJson.csproj" />
-    <ProjectReference Include="..\PPTail.Data.PhotoBlog\PPTail.Data.PhotoBlog.csproj" />
-    <ProjectReference Include="..\PPTail.Data.WordpressFiles\PPTail.Data.WordpressFiles.csproj" />
-    <ProjectReference Include="..\PPTail.Io\PPTail.Io.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\PPTail.Common\PPTail.Common.csproj" />
+		<ProjectReference Include="..\PPTail.Data.Ef\PPTail.Data.Ef.csproj" />
+		<ProjectReference Include="..\PPTail.Data.FileSystem\PPTail.Data.FileSystem.csproj" />
+		<ProjectReference Include="..\PPTail.Data.Forestry\PPTail.Data.Forestry.csproj" />
+		<ProjectReference Include="..\PPTail.Data.MediaBlog\PPTail.Data.MediaBlog.csproj" />
+		<ProjectReference Include="..\PPTail.Data.NativeJson\PPTail.Data.NativeJson.csproj" />
+		<ProjectReference Include="..\PPTail.Data.PhotoBlog\PPTail.Data.PhotoBlog.csproj" />
+		<ProjectReference Include="..\PPTail.Data.WordpressFiles\PPTail.Data.WordpressFiles.csproj" />
+		<ProjectReference Include="..\PPTail.Io\PPTail.Io.csproj" />
+	</ItemGroup>
 
 </Project>
