@@ -7,11 +7,13 @@ namespace PPTail.Entities
 {
     public class SiteSettings
     {
+        // TODO: Review which properties should be nullable
+
         public String Title { get; set; }
         public String Description { get; set; }
         public Int32 PostsPerPage { get; set; }
         public Int32 PostsPerFeed { get; set; }
-        public String Theme { get; set; } = string.Empty;
+        public String? Theme { get; set; }
         public String Copyright { get; set; }
         public String ContactEmail { get; set; }
         public bool UseAdditionalPagesDropdown { get; set; } = true;
