@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
-using PPTail;
+using PPTail.Console.Common.Extensions;
 using TestHelperExtensions;
 
-namespace PPTail.Test
+namespace PPTail.Console.Common.Test
 {
     [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
     public class StringExtensions_ValidateArguments_Should
@@ -14,8 +14,8 @@ namespace PPTail.Test
         [Fact]
         public void ReturnInvalidIfArgsIsNull()
         {
-            string[] target = null;
-            (bool isValid, var errors) = target.ValidateParameters();
+            string[]? target = null;
+            (bool isValid, var errors) = target!.ValidateParameters("PPTail.Console.Common.Test");
             Assert.False(isValid);
         }
 
@@ -26,7 +26,7 @@ namespace PPTail.Test
             String expectedTarget = string.Empty.GetRandom();
             String expectedTemplate = string.Empty.GetRandom();
             var target = new string[] { expectedSource, expectedTarget, expectedTemplate };
-            (bool isValid, var errors) = target.ValidateParameters();
+            (bool isValid, var errors) = target.ValidateParameters("PPTail.Console.Common.Test");
             Assert.True(isValid);
         }
 
@@ -37,7 +37,7 @@ namespace PPTail.Test
             String expectedTarget = string.Empty.GetRandom();
             String expectedTemplate = string.Empty.GetRandom();
             var target = new string[] { expectedSource, expectedTarget, expectedTemplate, expectedSource };
-            (bool isValid, var errors) = target.ValidateParameters();
+            (bool isValid, var errors) = target.ValidateParameters("PPTail.Console.Common.Test");
             Assert.False(isValid);
         }
 
@@ -48,7 +48,7 @@ namespace PPTail.Test
             String expectedTarget = string.Empty.GetRandom();
             String expectedTemplate = string.Empty.GetRandom();
             var target = new string[] { expectedSource, expectedTarget, expectedTemplate };
-            (bool isValid, var errors) = target.ValidateParameters();
+            (bool isValid, var errors) = target.ValidateParameters("PPTail.Console.Common.Test");
             Assert.False(isValid);
         }
 
@@ -59,7 +59,7 @@ namespace PPTail.Test
             String expectedTarget = string.Empty;
             String expectedTemplate = string.Empty.GetRandom();
             var target = new string[] { expectedSource, expectedTarget, expectedTemplate };
-            (bool isValid, var errors) = target.ValidateParameters();
+            (bool isValid, var errors) = target.ValidateParameters("PPTail.Console.Common.Test");
             Assert.False(isValid);
         }
 
@@ -70,7 +70,7 @@ namespace PPTail.Test
             String expectedTarget = string.Empty.GetRandom();
             String expectedTemplate = string.Empty;
             var target = new string[] { expectedSource, expectedTarget, expectedTemplate };
-            (bool isValid, var errors) = target.ValidateParameters();
+            (bool isValid, var errors) = target.ValidateParameters("PPTail.Console.Common.Test");
             Assert.False(isValid);
         }
 
