@@ -32,7 +32,7 @@ namespace PPTail.Data.FileSystem.Test
             var target = (null as IContentRepository).Create(fileProvider, directoryProvider.Object, rootPath);
             var actual = target.GetFolderContents(relativePath);
 
-            directoryProvider.Verify(fs => fs.EnumerateFiles(folderPath), Times.Once);
+            directoryProvider.Verify(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>()), Times.Once);
         }
 
         [Fact]
@@ -46,7 +46,7 @@ namespace PPTail.Data.FileSystem.Test
             var files = (null as IEnumerable<SourceFile>).Create(relativePath, expected);
 
             var directoryProvider = new Mock<IDirectory>();
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath)).Returns(files.Select(f => f.FileName));
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>())).Returns(files.Select(f => f.FileName));
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
 
             var fileProvider = new Mock<IFile>();
@@ -74,7 +74,7 @@ namespace PPTail.Data.FileSystem.Test
 
             var directoryProvider = new Mock<IDirectory>();
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
-            directoryProvider.Setup(dp => dp.EnumerateFiles(It.IsAny<string>()))
+            directoryProvider.Setup(dp => dp.EnumerateFiles(It.IsAny<string>(), It.IsAny<bool>()))
                 .Returns(files.Select(f => System.IO.Path.Combine(rootPath, relativePath, f.FileName)));
 
             var fileProvider = new Mock<IFile>();
@@ -100,7 +100,7 @@ namespace PPTail.Data.FileSystem.Test
 
             var directoryProvider = new Mock<IDirectory>();
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath)).Returns(files.Select(f => f.FileName));
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>())).Returns(files.Select(f => f.FileName));
 
             var fileProvider = new Mock<IFile>();
             foreach (var file in files)
@@ -127,7 +127,7 @@ namespace PPTail.Data.FileSystem.Test
 
             var directoryProvider = new Mock<IDirectory>();
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath)).Returns(files.Select(f => f.FileName));
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>())).Returns(files.Select(f => f.FileName));
 
             var fileProvider = new Mock<IFile>();
             foreach (var file in files)
@@ -155,7 +155,7 @@ namespace PPTail.Data.FileSystem.Test
 
             var directoryProvider = new Mock<IDirectory>();
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath)).Returns(files.Select(f => System.IO.Path.Combine(folderPath, f.FileName)));
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>())).Returns(files.Select(f => System.IO.Path.Combine(folderPath, f.FileName)));
 
             var fileProvider = new Mock<IFile>();
             foreach (var file in files)
@@ -180,7 +180,7 @@ namespace PPTail.Data.FileSystem.Test
 
             var directoryProvider = new Mock<IDirectory>();
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(false);
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath))
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>()))
                 .Throws(new System.IO.DirectoryNotFoundException());
 
             var target = (null as IContentRepository).Create(Mock.Of<IFile>(), directoryProvider.Object, rootPath);
@@ -202,7 +202,7 @@ namespace PPTail.Data.FileSystem.Test
             fileNames.AddRange(files.Select(f => System.IO.Path.Combine(folderPath, f.FileName)));
 
             var directoryProvider = new Mock<IDirectory>();
-            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath)).Returns(fileNames);
+            directoryProvider.Setup(fs => fs.EnumerateFiles(folderPath, It.IsAny<bool>())).Returns(fileNames);
             directoryProvider.Setup(fs => fs.Exists(folderPath)).Returns(true);
 
             var fileProvider = new Mock<IFile>();
