@@ -62,17 +62,17 @@ namespace PPTail.SiteGenerator
             var categories = contentRepo.GetCategories();
 
             // Create Sidebar Content
-            if (logger.IsNotNull()) logger.LogInformation("Generating sidebar content");
+            if (logger is not null) logger.LogInformation("Generating sidebar content");
             var rootLevelSidebarContent = pageGen.GenerateSidebarContent(posts, pages, widgets, "./");
             var childLevelSidebarContent = pageGen.GenerateSidebarContent(posts, pages, widgets, "../");
 
             // Create navbars
-            if (logger.IsNotNull()) logger.LogInformation("Generating navbar content");
+            if (logger is not null) logger.LogInformation("Generating navbar content");
             var rootLevelNavigationContent = navProvider.CreateNavigation(pages, "./", siteSettings.OutputFileExtension);
             var childLevelNavigationContent = navProvider.CreateNavigation(pages, "../", siteSettings.OutputFileExtension);
 
             // Create bootstrap file
-            if (logger.IsNotNull()) logger.LogInformation("Generating bootstrap file");
+            if (logger is not null) logger.LogInformation("Generating bootstrap file");
             var bootstrapFile = new SiteFile()
             {
                 RelativeFilePath = $"./bootstrap.min.css",
@@ -83,7 +83,7 @@ namespace PPTail.SiteGenerator
                 result.Add(bootstrapFile);
 
             // Create Style page
-            if (logger.IsNotNull()) logger.LogInformation("Generating style (css) file");
+            if (logger is not null) logger.LogInformation("Generating style (css) file");
             result.Add(new SiteFile()
             {
                 RelativeFilePath = $"./Style.css",
@@ -92,7 +92,7 @@ namespace PPTail.SiteGenerator
             });
 
             // Create home page
-            if (logger.IsNotNull()) logger.LogInformation("Generating home page");
+            if (logger is not null) logger.LogInformation("Generating home page");
             result.Add(new SiteFile()
             {
                 RelativeFilePath = $"./index.html",
@@ -101,7 +101,7 @@ namespace PPTail.SiteGenerator
             });
 
             // Create Archive
-            if (logger.IsNotNull()) logger.LogInformation("Generating archive page");
+            if (logger is not null) logger.LogInformation("Generating archive page");
             result.Add(new SiteFile()
             {
                 RelativeFilePath = $"./archive.html",
@@ -110,7 +110,7 @@ namespace PPTail.SiteGenerator
             });
 
             // Create Contact Page
-            if (logger.IsNotNull()) logger.LogInformation("Generating contact page");
+            if (logger is not null) logger.LogInformation("Generating contact page");
             result.Add(new SiteFile()
             {
                 RelativeFilePath = $"./contact.html",
@@ -119,7 +119,7 @@ namespace PPTail.SiteGenerator
             });
 
             // Create RSS Feed
-            if (logger.IsNotNull()) logger.LogInformation("Generating RSS feed");
+            if (logger is not null) logger.LogInformation("Generating RSS feed");
             var syndicationContent = syndicationProvider.GenerateFeed(posts);
             result.Add(new SiteFile()
             {
@@ -128,7 +128,7 @@ namespace PPTail.SiteGenerator
                 Content = syndicationContent
             });
 
-            if (logger.IsNotNull()) logger.LogInformation("Generating post pages");
+            if (logger is not null) logger.LogInformation("Generating post pages");
             foreach (var post in posts)
             {
                 // Add all published content pages to the results
@@ -141,7 +141,7 @@ namespace PPTail.SiteGenerator
                     // Add the post page
                     String postFilePath = System.IO.Path.Combine("Posts", postFileName);
                     var postPageTemplateType = Enumerations.TemplateType.PostPage;
-                    if (logger.IsNotNull()) logger.LogInformation("Generating post page for PostId: {PostId}", post.Id);
+                    if (logger is not null) logger.LogInformation("Generating post page for PostId: {PostId}", post.Id);
                     result.Add(new SiteFile()
                     {
                         RelativeFilePath = postFilePath,
@@ -156,7 +156,7 @@ namespace PPTail.SiteGenerator
                     String permalinkFileName = $"{contentEncoder.HTMLEncode(post.Id.ToString())}.{siteSettings.OutputFileExtension}";
                     String permalinkFilePath = System.IO.Path.Combine("Permalinks", permalinkFileName);
                     String redirectFilePath = $"../Posts/{postFileName}"; // System.IO.Path.Combine("..", postFilePath);
-                    if (logger.IsNotNull()) logger.LogInformation("Publishing post page for PostId: {PostId}", post.Id);
+                    if (logger is not null) logger.LogInformation("Publishing post page for PostId: {PostId}", post.Id);
                     result.Add(new SiteFile()
                     {
                         RelativeFilePath = permalinkFilePath,
@@ -166,13 +166,13 @@ namespace PPTail.SiteGenerator
                 }
             }
 
-            if (logger.IsNotNull()) logger.LogInformation("Generating content pages");
+            if (logger is not null) logger.LogInformation("Generating content pages");
             foreach (var page in pages)
             {
                 // Add all published content pages to the results
                 if (page.IsPublished)
                 {
-                    if (logger.IsNotNull()) logger.LogInformation("Generating content page for Id: {Id}", page.Id);
+                    if (logger is not null) logger.LogInformation("Generating content page for Id: {Id}", page.Id);
 
                     if (string.IsNullOrWhiteSpace(page.Slug))
                         page.Slug = contentEncoder.UrlEncode(page.Title);
@@ -188,7 +188,7 @@ namespace PPTail.SiteGenerator
             }
 
             // Add Search Pages
-            if (logger.IsNotNull()) logger.LogInformation("Generating search pages");
+            if (logger is not null) logger.LogInformation("Generating search pages");
             var tags = posts.GetAllTags().Distinct();
 
             var categoryIds = posts.SelectMany(p => p.CategoryIds).Distinct();
@@ -218,10 +218,10 @@ namespace PPTail.SiteGenerator
             }
 
             // Add favicon.ico file if it exists
-            if (logger.IsNotNull()) logger.LogInformation("Add favicon.ico file");
+            if (logger is not null) logger.LogInformation("Add favicon.ico file");
             var rootFiles = contentRepo.GetFolderContents(".");
             var faviconFile = rootFiles.SingleOrDefault(f => f.FileName == "favicon.ico");
-            if (faviconFile.IsNotNull())
+            if (faviconFile is not null)
             {
                 result.Add(new SiteFile()
                 {
@@ -233,7 +233,7 @@ namespace PPTail.SiteGenerator
             }
 
             // Add files from Theme if there are any
-            if (logger.IsNotNull()) logger.LogInformation("Add files from theme");
+            if (logger is not null) logger.LogInformation("Add files from theme");
             if (!string.IsNullOrWhiteSpace(siteSettings.Theme))
             {
                 String path = $".\\themes\\{siteSettings.Theme}";
@@ -251,8 +251,8 @@ namespace PPTail.SiteGenerator
             }
 
             // Add additional raw files
-            if (logger.IsNotNull()) logger.LogInformation("Add additional raw files");
-            if (siteSettings.AdditionalFilePaths.IsNotNull() && siteSettings.AdditionalFilePaths.Any())
+            if (logger is not null) logger.LogInformation("Add additional raw files");
+            if (siteSettings.AdditionalFilePaths is not null && siteSettings.AdditionalFilePaths.Any())
             {
                 var additionalFiles = contentRepo.GetFoldersContents(siteSettings.AdditionalFilePaths, true);
                 foreach (var rawFile in additionalFiles)
